Not scrolling back when new content arrives
By default, when new content arrives, the bottomInset resets; so the
scrollView scrolls up a bit and the only notification for new content is
decrease length of scroll bar.
I've changed this so the bottomInset won't reset and there is always a
constant inset at the bottom of table, except when the scrollable
content has been ended.